space/requirements.txt
torch
diffusers==0.32.2
transformers==4.47.1
accelerate
Pillow
